# baseline/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Education Agent Copilot Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/copilot/run/confirm": {
      "post": {
        "tags": [
          "copilot"
        ],
        "summary": "Confirm Study Pilot",
        "operationId": "confirm_study_pilot_api_v1_copilot_run_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyPilotClarificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyPilotRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/copilot/run/expand-blocks": {
      "post": {
        "tags": [
          "copilot"
        ],
        "summary": "Expand Block Plan",
        "operationId": "expand_block_plan_api_v1_copilot_run_expand_blocks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockPlanExpandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyPilotRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/copilot/plan/save": {
      "post": {
        "tags": [
          "copilot"
        ],
        "summary": "Save Plan",
        "operationId": "save_plan_api_v1_copilot_plan_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/copilot/plan/current": {
      "get": {
        "tags": [
          "copilot"
        ],
        "summary": "Get Current Plan",
        "operationId": "get_current_plan_api_v1_copilot_plan_current_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/copilot/plan/progress": {
      "post": {
        "tags": [
          "copilot"
        ],
        "summary": "Update Plan Progress",
        "operationId": "update_plan_progress_api_v1_copilot_plan_progress_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanProgressUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_dict_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "ok"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "ApiResponse[dict]"
      },
...
      "BlockPlan": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "day": {
            "type": "string",
            "title": "Day"
          },
          "limitMinutes": {
            "type": "integer",
            "title": "Limitminutes"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/BlockItem"
            },
            "type": "array",
            "title": "Blocks"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "day",
          "limitMinutes",
          "blocks"
        ],
        "title": "BlockPlan"
      },
      "BlockPlanExpandRequest": {
        "properties": {
          "normalized": {
            "$ref": "#/components/schemas/StudyPlanRequest"
          },
          "blockPlan": {
            "$ref": "#/components/schemas/BlockPlan"
          }
        },
        "type": "object",
        "required": [
          "normalized",
          "blockPlan"
        ],
        "title": "BlockPlanExpandRequest"
      },
...
      "ClarificationAnswer": {
        "properties": {
          "questionId": {
            "type": "string",
            "title": "Questionid"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "questionId",
          "answer"
        ],
        "title": "ClarificationAnswer"
      },
      "ClarificationPrompt": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/ClarificationQuestion"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "title",
          "description",
          "questions"
        ],
        "title": "ClarificationPrompt"
      },
...
      "FrontendMemory": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": ""
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "content": {
            "type": "string",
            "title": "Content",
            "default": ""
          }
        },
        "type": "object",
        "title": "FrontendMemory"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanProgressUpdateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "default"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id",
            "default": ""
          },
          "plan_id": {
            "type": "string",
            "title": "Plan Id",
            "default": ""
          },
          "plan_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Version"
          },
          "task_key": {
            "type": "string",
            "title": "Task Key"
          },
          "done": {
            "type": "boolean",
            "title": "Done"
          },
          "task_title": {
            "type": "string",
            "title": "Task Title",
            "default": ""
          },
          "task_type": {
            "type": "string",
            "title": "Task Type",
            "default": ""
          },
          "actual_minutes": {
            "type": "integer",
            "title": "Actual Minutes",
            "default": 0
          },
          "plan_message": {
            "type": "string",
            "title": "Plan Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "task_key",
          "done"
        ],
        "title": "PlanProgressUpdateRequest"
      },
...
      "SavePlanRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": "default"
          },
          "plan": {
            "additionalProperties": true,
            "type": "object",
            "title": "Plan"
          }
        },
        "type": "object",
        "required": [
          "plan"
        ],
        "title": "SavePlanRequest"
      },
...
      "StudyPilotClarificationRequest": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/ClarificationAnswer"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "sessionId"
        ],
        "title": "StudyPilotClarificationRequest"
      },
...
      "StudyPilotRunResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "maintenance",
              "needs_clarification"
            ],
            "title": "Status",
            "default": "ready"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "backend-live"
          },
          "request": {
            "additionalProperties": true,
            "type": "object",
            "title": "Request"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "followUp": {
            "type": "string",
            "title": "Followup"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "next_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next Steps"
          },
          "clarification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClarificationPrompt"
              },
              {
                "type": "null"
              }
            ]
          },
          "blockPlan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlockPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StudyPlanPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "memory_used": {
            "items": {
              "$ref": "#/components/schemas/FrontendMemory"
            },
            "type": "array",
            "title": "Memory Used"
          },
          "memory_candidates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Memory Candidates"
          },
          "model_provider": {
            "type": "string",
            "title": "Model Provider",
            "default": ""
          },
          "model_used": {
            "type": "string",
            "title": "Model Used",
            "default": ""
          },
          "is_fallback": {
            "type": "boolean",
            "title": "Is Fallback",
            "default": false
          },
          "error_message": {
            "type": "string",
            "title": "Error Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "request",
          "message",
          "followUp"
        ],
        "title": "StudyPilotRunResponse"
      },
      "StudyPlanPayload": {
        "properties": {
          "learner_profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Learner Profile"
          },
          "retrieved_context": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Retrieved Context"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStageResult"
            },
            "type": "array",
            "title": "Stages"
          },
          "weekly_plan": {
            "items": {
              "$ref": "#/components/schemas/StudyDayPlan"
            },
            "type": "array",
            "title": "Weekly Plan"
          },
          "final_message": {
            "type": "string",
            "title": "Final Message"
          },
          "next_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next Actions"
          }
        },
        "type": "object",
        "required": [
          "learner_profile",
          "retrieved_context",
          "stages",
          "weekly_plan",
          "final_message",
          "next_actions"
        ],
        "title": "StudyPlanPayload"
      },
...
    }
  }
}